# backend/swagger.yaml
swagger: "2.0"
info:
  description: "Trainer REST API documentation"
  version: "1.0.0"
  title: "Trainer"
host: "trainer.ksi.fit.cvut.cz"
basePath: "/api"
...
schemes:
- "https"
paths:
...
  /courses:
    get:
      tags:
        - "courses"
      summary: "Get course list"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Success, course list"
          schema:
            $ref: "#/definitions/CourseFindDTOArray"
      security:
        - login_secret: [ ]
...
  /courses/{id}/secret:
    put:
      tags:
        - "courses"
      summary: "Set course secret"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Unique identifier of the course"
          type: "integer"
          required: true
        - in: "body"
          name: "secret"
          required: true
          schema:
            $ref: "#/definitions/CourseSecretDTO"
      responses:
        "200":
          description: "Success, secret updated"
          schema:
            $ref: "#/definitions/CourseGetDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to set course secret"
        "404":
          description: "Course with given identifier not found"
      security:
        - login_secret: [ ]
  /courses/me:
    put:
      tags:
        - "courses"
      summary: "Join course with given secret"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "secret"
          required: true
          schema:
            $ref: "#/definitions/CourseSecretDTO"
      responses:
        "200":
          description: "Success, course joined"
          schema:
            $ref: "#/definitions/CourseGetDTO"
        "401":
          description: "Unauthorized user"
        "400":
          description: "User is already member of the course"
        "404":
          description: "Course with given secret not found"
      security:
        - login_secret: [ ]
  /courses/{id}/weeks/{weekId}:
    get:
      tags:
        - "courses"
      summary: "Get description of week in course with given id"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique course identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "path"
          description: "Unique week identifier"
          name: "weekId"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, week detail (⚠️ without lessons ⚠️)"
          schema:
            $ref: "#/definitions/WeekFindDTO"
      security:
        - login_secret: [ ]

  # Course Users
...
  /lessons:
    post:
      tags:
        - "lessons"
      summary: "Create new lesson"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Lesson to be created"
          required: true
          schema:
            $ref: "#/definitions/LessonCreateDTO"
      responses:
        "201":
          description: "Success, lesson created"
          schema:
            $ref: "#/definitions/LessonGetDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to create new lesson (not a teacher)"
      security:
        - login_secret: []
  /lessons/current:
    get:
      tags:
        - "lessons"
      summary: "Get current running lessons"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Success, current lesson list"
          schema:
            $ref: "#/definitions/LessonGetDTOArray"
      security:
        - login_secret: [ ]
...
  /lessons/{id}/editData:
    get:
      tags:
        - "lessons"
      summary: "Get detail of lesson with given id without user data (teacher only)"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, lesson detail (without user data)"
          schema:
            $ref: "#/definitions/LessonGetDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights (not a teacher in the lesson)"
        "404":
          description: "Lesson with given identifier not found"
      security:
        - login_secret: [ ]
  /lessons/{id}/week:
    get:
      tags:
        - "lessons"
      summary: "Get detail of lesson week (with all lessons and modules)"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, lesson week detail"
          schema:
            $ref: "#/definitions/WeekDetailDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to see lesson"
        "404":
          description: "Lesson with given identifier not found"
      security:
        - login_secret: [ ]
  /lessons/{id}/courses/{courseId}:
    post:
      tags:
        - "lessons"
      summary: "Copy lesson (with all of its modules) into another course"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "path"
          description: "Identifier of course to copy into"
          name: "courseId"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, copied lesson detail"
          schema:
            $ref: "#/definitions/LessonGetDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to edit original lesson / new course"
        "404":
          description: "Lesson or course with given identifier not found"
      security:
        - login_secret: [ ]
  /lessons/{id}/code:
    post:
      tags:
        - "lessons"
      summary: "Unlock all modules in given lesson"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "DTO with unlock code"
          required: true
          schema:
            $ref: "#/definitions/LessonUnlockDTO"
      responses:
        "200":
          description: "Success, lesson modules unlocked"
        "400":
          description: "Failure, wrong code"
        "401":
          description: "Unauthorized user"
        "404":
          description: "Lesson with given identifier not found"
      security:
        - login_secret: [ ]

  # Lesson users
...
securityDefinitions:
  login_secret:
    type: "apiKey"
    name: "LoginSecret"
    in: "header"
definitions:
...
  CourseFindDTOArray:
    type: "array"
    items:
      $ref: "#/definitions/CourseFindDTO"
  CourseFindDTO:
    description: "Course list DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique course identifier"
        example: 1
      name:
        type: "string"
        description: "Course name"
        example: "Programming and Algorithms 1 (3 + 4)"
      subject:
        $ref: "#/definitions/SubjectFindDTO"
      semester:
        $ref: "#/definitions/SemesterFindDTO"
      lessonCompleted:
        type: "integer"
        description: "Number of completed lessons in course"
        example: 2
      lessonCount:
        type: "integer"
        description: "Number of lessons in course"
        example: 5
      role:
        $ref: "#/definitions/RoleEnum"
  CourseGetDTO:
    description: "Course get DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique course identifier"
        example: 1
      name:
        type: "string"
        description: "Course name"
        example: "Programming and Algorithms 1 (3 + 4)"
      subject:
        $ref: "#/definitions/SubjectFindDTO"
      semester:
        $ref: "#/definitions/SemesterFindDTO"
      role:
        $ref: "#/definitions/RoleEnum"
      secret:
        type: "string"
        description: "Course join secret (only shown to teachers)"
        example: null
      weeks:
        $ref: "#/definitions/WeekGetDTOArray"
  CourseSecretDTO:
    description: "Course secret"
    type: "object"
    properties:
      secret:
        type: "string"
        description: "Secret for joining course"
        example: "730NE"
  SubjectFindDTO:
    description: "Subject DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique subject identifier"
        example: 1
      name:
        type: "string"
        description: "Subject name"
        example: "Programming and Algorithms 1"
      code:
        type: "string"
        description: "Subject code"
        example: "BI-PA1.21"
  SemesterFindDTO:
    description: "Semester DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique semester identifier"
        example: 1
      code:
        type: "string"
        description: "Semester code"
        example: "B232"
      from:
        type: "string"
        description: "Semester start date"
        format: "date-time"
      until:
        type: "string"
        description: "Semester end date"
        format: "date-time"

  # Week
  WeekFindDTO:
    description: "Week"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique week identifier"
        example: 1
      name:
        type: "string"
        description: "Week name"
        example: "1st week"
      from:
        type: "string"
        format: "date-time"
        description: "Week start date"
        example: "2024-02-01T00:00:00.000+00:00"
      until:
        type: "string"
        format: "date-time"
        description: "Week end date"
        example: "2025-02-01T00:00:00.000+00:00"
      course:
        $ref: "#/definitions/CourseFindDTO"
  WeekGetDTOArray:
    type: "array"
    items:
      $ref: "#/definitions/WeekGetDTO"
  WeekGetDTO:
    description: "Week"
    allOf:
      - $ref: "#/definitions/WeekFindDTO"
    properties:
      lessons:
        $ref: "#/definitions/LessonFindDTOArray"
  WeekDetailDTO:
    description: "Week"
    allOf:
      - $ref: "#/definitions/WeekFindDTO"
    properties:
      lessons:
        $ref: "#/definitions/LessonGetDTOArray"
...
  CourseUserReadDTO:
    description: "Course user"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique user identifier"
        example: 1
      username:
        type: "string"
        description: "User's username"
        example: "wrzecond"
      name:
        type: "string"
        description: "User's name"
        example: "Ondřej Wrzecionko"
      role:
        $ref: "#/definitions/RoleEnum"
      progress:
        type: "integer"
        description: "User progress in course (in percent)"
        example: 24
...
  LessonDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique lesson identifier"
        example: 1
      name:
        type: "string"
        description: "Lesson name"
        example: "Compilation"
      hidden:
        type: "boolean"
        description: "Is lesson hidden from students?"
        example: false
      type:
        $ref: "#/definitions/LessonType"
      lockCode:
        type: "string"
        description: "Code for unlocking modules in lesson (teacher only)"
        example: null
      timeLimit:
        type: "string"
        format: "date-time"
        description: "Lesson end date"
        example: "2025-02-01T00:00:00.000+00:00"
  LessonFindDTOArray:
    type: "array"
    items:
      $ref: "#/definitions/LessonFindDTO"
  LessonFindDTO:
    description: "Lesson"
    type: "object"
    allOf:
      - $ref: "#/definitions/LessonDTO"
    properties:
      order:
        type: "integer"
        description: "Lesson order in week"
        example: 1
      progress:
        type: "integer"
        description: "Overall progress in lesson"
        example: 50
  LessonGetDTOArray:
    type: "array"
    items:
      $ref: "#/definitions/LessonGetDTO"
  LessonGetDTO:
    description: "Lesson"
    type: "object"
    allOf:
      - $ref: "#/definitions/LessonDTO"
    properties:
      week:
        $ref: "#/definitions/WeekFindDTO"
      timeStart:
        type: "string"
        format: "date-time"
        description: "Lesson start date"
        example: "2024-02-01T00:00:00.000+00:00"
      description:
        type: "string"
        description: "Lesson description (introduction text)"
        example: "In this lesson, we will cover the basics of how to compile..."
      modules:
        $ref: "#/definitions/ModuleFindDTOArray"
  LessonType:
    type: "string"
    enum: [ "TUTORIAL", "TUTORIAL_PREPARATION", "INDIVIDUAL_TASK", "SUPPLEMENTARY", "FREQUENT_MISTAKES", "INFORMATION" ]
    description: "Lesson type"
    example: "TUTORIAL"
  LessonUpdateDTO:
    type: "object"
    description: "DTO used for updating lesson"
    properties:
      name:
        type: "string"
        description: "Newly created lesson name"
        example: "Another lesson"
      hidden:
        type: "boolean"
        description: "Should lesson be hidden from students?"
        example: false
      order:
        type: "integer"
        description: "Lesson order in week"
        example: 2
      type:
        $ref: "#/definitions/LessonType"
      lockCode:
        type: "string"
        description: "Code used for unlocking lesson modules"
        example: "secretcode"
      timeStart:
        type: "string"
        format: "date-time"
        description: "Lesson start date"
        example: "2024-02-01T00:00:00.000+00:00"
      timeLimit:
        type: "string"
        format: "date-time"
        description: "Time limit for user to solve lesson in time"
        example: "2025-02-01T00:00:00.000+00:00"
      description:
        type: "string"
        description: "Lesson description (with assignment, conditions...)"
        example: "In this lesson, we will cover the basics of how to compile..."
  LessonCreateDTO:
    allOf:
      - $ref: "#/definitions/LessonUpdateDTO"
    properties:
      weekId:
        type: "integer"
        description: "Identifier of week for lesson to be added to"
        example: 1
  LessonUnlockDTO:
    type: "object"
    description: "DTO used for unlocking lesson"
    properties:
      code:
        type: "string"
        description: "Code to set (null for unsetting)"
        example: "secretcode"

  # Lesson user
...
  ModuleFindDTOArray:
    type: "array"
    items:
      $ref: "#/definitions/ModuleFindDTO"
  ModuleFindDTO:
    allOf:
      - $ref: "#/definitions/ModuleGetDTO"
    properties:
      locked:
        type: "boolean"
        description: "Flag determining, whether module is really locked"
        example: false
      file:
        type: "string"
        description: "If module has file (test data), it will be the name of the file"
        example: "d32kvcz21y"
      author:
        type: "string"
        description: "Module author"
        example: "wrzecond"
      editors:
        type: "array"
        description: "Module editors (usernames)"
        items:
          type: "string"
        example: ["matouj10", "pasekmat"]
      allowedShow:
        type: "boolean"
        description: "Flag determining whether student allowed anonymized publishing of the module"
        example: false
      completedEarly:
        type: "boolean"
        description: "Flag determining whether student completed the module before lesson end"
        example: true
      completed:
        type: "string"
        format: "date-time"
        description: "Datetime when student completed the module (null if not completed)"
        example: "2024-03-01T14:16:11.000+00:00"
      requestType:
        type: "string"
        enum: ["HELP", "EVALUATE"]
        description: "Type of request that is currently active"
        example: "EVALUATE"
      teacherComment:
        type: "string"
        description: "Teacher's comment on the last student request"
        example: "It looks good, but try better."
      requestText:
        type: "string"
        description: "Student's comment associated with current active request"
        example: "I improved it, please evaluate it."
      order:
        type: "integer"
        description: "Module order in lesson"
        example: 2
      progress:
        type: "integer"
        description: "Student progress in module (in percent)"
        example: 100
      hasStudentFile:
        type: "boolean"
        description: "Flag indicating whether module has a solution (file) uploaded by student"
        example: true
  ModuleGetDTO:
    description: "Module"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique module identifier"
        example: 1
      name:
        type: "string"
        description: "Module name"
        example: "Assignment: Compile on your own computer"
      depends:
        type: "integer"
        description: "ID of module this module is depending on"
        example: 2
      author:
        $ref: "#/definitions/CourseUserReadDTO"
      editors:
        type: "array"
        description: "IDs of users who are editors of the module"
        items:
          type: "integer"
      type:
        type: "string"
        enum: ["TEXT", "QUIZ", "CODE", "ASSIGNMENT"]
        description: "Type of module"
        example: "CODE"
      lastModificationTime:
        type: "string"
        format: "date-time"
        description: "Last time, when module was changed"
        example: "2024-02-13T11:03:20.000+00:00"
      difficulty:
        type: "string"
        enum: ["BEGINNER", "EASY", "MEDIUM", "DIFFICULT", "EXTREME"]
        description: "Module difficulty rating"
        example: "EASY"
      assignment:
        type: "string"
        description: "Markdown assignment of module"
        example: "In this task, you have to compile the following code on your computer and upload it."
      minPercent:
        type: "integer"
        description: "Minimum percent needed for completing the module"
        example: 100
      lockable:
        type: "boolean"
        description: "Flag determining, whether module can be locked."
        example: false
      timeLimit:
        type: "boolean"
        description: "Flag determining, whether module should be solved 'early'."
        example: true
      manualEval:
        type: "boolean"
        description: "Flag determining, whether module should be evaluated manually (by teacher)."
        example: true
      hidden:
        type: "boolean"
        description: "Flag determining, whether module is hidden from students."
        example: false
...
  RoleEnum:
    type: "string"
    description: "User role in course"
    enum: ["TEACHER", "STUDENT"]
    example: "STUDENT"

  # User
...